requirements-dev.txt
# Core dependencies
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.5.0
torchvision==0.20.0
numpy>=1.24.0
Pillow>=10.0.0
runpod

# Vercel Blob Storage
git+https://github.com/timpietrusky/vercel_blob.git

# Testing tools
pytest==8.0.0
pytest-assume==2.4.3

# Code quality tools
ruff==0.3.0  # Fast all-in-one Python linter
pre-commit==3.6.2  # Git pre-commit hooks
python-dotenv==1.0.1  # for loading environment variables